refactor: move span batch type into separate module
