+++ src/derive/stages/mod.rs
@@ -2,3 +2,4 @@ pub mod attributes;
 pub mod batcher_transactions;
 pub mod batches;
 pub mod channels;
+mod span_batch;
